+++ cypress/integration/Platform/44007_spec.js
@@ -26,13 +26,10 @@ describe("Platfrom 44007", () => {
   it("Shows current conditions", () => {
     cy.visit(platformUrl)
 
-    cy.contains("Current Conditions")
+    cy.contains("Latest Conditions")
     cy.contains("Winds -")
 
-    cy.get("[style='margin-top: 1rem;'] > :nth-child(2) > .row")
-      .children()
-      .its("length")
-      .should("be.gte", 4)
+    cy.get("[style='margin-top: 1rem;'] > :nth-child(2) > .row").children().its("length").should("be.gte", 4)
 
     cy.contains("Knots")
   })
@@ -41,20 +38,12 @@ describe("Platfrom 44007", () => {
     cy.visit(platformUrl)
 
     cy.contains("Observations").click()
-    cy.get("[href='/platform/44007/observations/wind']")
-      .first()
-      .click()
+    cy.get("[href='/platform/44007/observations/wind']").first().click()
     cy.get("h4").contains("Wind")
     cy.get("svg.highcharts-root")
-    cy.get("svg.highcharts-root")
-      .contains("Gust")
-      .click()
-    cy.get("svg.highcharts-root")
-      .contains("Speed")
-      .click()
-    cy.get("svg.highcharts-root")
-      .contains("Direction")
-      .click()
+    cy.get("svg.highcharts-root").contains("Gust").click()
+    cy.get("svg.highcharts-root").contains("Speed").click()
+    cy.get("svg.highcharts-root").contains("Direction").click()
     cy.get("svg.highcharts-root").contains("Knots")
   })
 
@@ -72,21 +61,14 @@ describe("Platfrom 44007", () => {
     cy.contains("English").click()
     cy.get("svg.highcharts-root").contains("Feet")
     // cy.contains("Significant Wave Height observed").click()
-    cy.get("svg.highcharts-root")
-      .contains("Bedford Institute Wave Model")
-      .click()
+    cy.get("svg.highcharts-root").contains("Bedford Institute Wave Model").click()
   })
 
   it("Has More info menu", () => {
     cy.visit(platformUrl)
 
     cy.contains("More info").click()
-    cy.contains("More info")
-      .parent()
-      .children()
-      .last()
-      .children()
-      .should("have.length", 10)
+    cy.contains("More info").parent().children().last().children().should("have.length", 10)
 
     cy.contains("All Data From This Station")
     cy.contains("Compare Stations")
@@ -102,7 +84,7 @@ describe("Platfrom 44007", () => {
   it("Updated recently", () => {
     cy.visit(platformUrl)
 
-    cy.contains("Last updated at:").then($element => {
+    cy.contains("Last updated at:").then(($element) => {
       const text = $element[0].parentElement.innerText
 
       const date = new Date(text.split("Last updated at: "))